+++ Assets/lilToon/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the sampler of the normal map 2nd to `Repeat`
 - Changed to move the calculation of point light, spot light and area light to the pixel shader in HDRP and also calculate the light direction
 - Adjusted GUI a little
+- Optimized build time
 
 ### Fixed
 - Fixed a error in gem shader's GUI